% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_help.R
\name{mv_help}
\alias{mv_help}
\title{Display macroverse help information}
\usage{
mv_help()
}
\value{
Invisibly returns NULL. Prints help information to the console.
}
\description{
Provides an overview of the macroverse ecosystem and its packages.
}
\examples{
mv_help()

}
